Added new type system check: Cannot pass little-endian types to big-endian parameters, and vice versa. Returning big-endian to little-endian on assignment ie `le32 x = foo();` where foo's rtype is big-endian is still fine and will auto-convert.

+++ tests/vaultd.fx
@@ -267,7 +267,7 @@ def main() -> int
     };
 
     // Connect to server
-    if (!client_socket.connect("127.0.0.1\0", (i16)8080))
+    if (!client_socket.connect("127.0.0.1\0", 8080))
     {
         print("Failed to connect to server\n\0");
         client_socket.close();
